sm8150-common: Adjust powerhint for QOS interface
Currently all writes to the dev interface are done
by WriteStringToFd that uses char internally.

The current values seem to expect to be written as s32,
this sadly won't happen and triggers a special case in
the write handling of pm_qos_power_write:
it expects char buffers to be encoded as base16 and
decodes them.
This means the current 44 is actually 0x44 -> 68 seen by
the kernel.

Luckily it seems like both accepted values for this node
don't hit the threshold to enter C2, so it was never
noticed in real usage and didn't effect the device
C-States handling during hints.

Signed-off-by: Luca Stefani <luca.stefani.ge1@gmail.com>
Change-Id: I4fe9f5e83cf4134dda1cf3e68c413e0440536baf